Save calculated lengths to Database
